Added functionality for J&S, C&L Annuities. Do not completely agree with SOA calculator regarding Beneficiary mortality rates, but do not have a better assumption on hand. Spent a few days sorting that out.
Note addition of code in calculateDiscountedPV which saves the final calc df to a dataframe. This is temporary while I build out all the core functionality of the system.

Need to refactor calculator to have only a single primary calc df which is operated on by all functions. This would be a global variable and I effing hate global variables (Or rather, I strongly prefer functions with return values over functions which modify global data structures). But it just makes more sense here.